+++ src/app.rs
@@ -5,6 +5,7 @@ use std::{error::Error as StdError, marker::PhantomData, path::Path, sync::Arc, 
 use log::Level;
 use rayon::ThreadPoolBuilder;
 use shred::Resource;
+
 #[cfg(feature = "profiler")]
 use thread_profiler::{register_thread_with_profiler, write_profile};
 use winit::Event;
@@ -23,7 +24,7 @@ use {
     },
     error::{Error, Result},
     game_data::DataInit,
-    state::{State, StateData, StateMachine},
+    state::{State, StateData, StateMachine, TransEvent},
     state_event::{StateEvent, StateEventReader},
     ui::UiEvent,
 };
@@ -41,7 +42,11 @@ use {
 /// - `R`: `EventReader` implementation for the given event type `E`
 #[derive(Derivative)]
 #[derivative(Debug)]
-pub struct CoreApplication<'a, T, E = StateEvent, R = StateEventReader> {
+pub struct CoreApplication<'a, T, E = StateEvent, R = StateEventReader>
+where
+    T: 'static,
+    E: 'static,
+{
     /// The world
     #[derivative(Debug = "ignore")]
     world: World,
@@ -50,6 +55,8 @@ pub struct CoreApplication<'a, T, E = StateEvent, R = StateEventReader> {
     #[derivative(Debug = "ignore")]
     events: Vec<E>,
     event_reader_id: ReaderId<Event>,
+    #[derivative(Debug = "ignore")]
+    trans_reader_id: ReaderId<TransEvent<T, E>>,
     states: StateMachine<'a, T, E>,
     ignore_window_close: bool,
     data: T,
@@ -130,6 +137,7 @@ pub type Application<'a, T> = CoreApplication<'a, T, StateEvent, StateEventReade
 
 impl<'a, T, E, R> CoreApplication<'a, T, E, R>
 where
+    T: 'static,
     E: Clone + Send + Sync + 'static,
 {
     /// Creates a new Application with the given initial game state.
@@ -195,7 +203,7 @@ where
     ///
     /// This is identical in function to
     /// [ApplicationBuilder::new](struct.ApplicationBuilder.html#method.new).
-    pub fn build<P, S>(path: P, initial_state: S) -> Result<ApplicationBuilder<S, E, R>>
+    pub fn build<P, S>(path: P, initial_state: S) -> Result<ApplicationBuilder<S, T, E, R>>
     where
         P: AsRef<Path>,
         S: State<T, E> + 'a,
@@ -294,6 +302,22 @@ where
             states.stop(StateData::new(world, &mut self.data));
         }
 
+        // Read the Trans queue and apply changes.
+        {
+            let mut world = &mut self.world;
+            let states = &mut self.states;
+            let reader = &mut self.trans_reader_id;
+
+            let trans = world
+                .read_resource::<EventChannel<TransEvent<T, E>>>()
+                .read(reader)
+                .map(|e| e())
+                .collect::<Vec<_>>();
+            for tr in trans {
+                states.transition(tr, StateData::new(&mut world, &mut self.data));
+            }
+        }
+
         {
             #[cfg(feature = "profiler")]
             profile_scope!("handle_event");
@@ -363,16 +387,19 @@ impl<'a, T, E, R> Drop for CoreApplication<'a, T, E, R> {
 /// using a custom set of configuration. This is the normal way an
 /// [`Application`](struct.Application.html)
 /// object is created.
-pub struct ApplicationBuilder<S, E, R> {
+pub struct ApplicationBuilder<S, T, E, R> {
     // config: Config,
     initial_state: S,
     /// Used by bundles to access the world directly
     pub world: World,
     ignore_window_close: bool,
-    phantom: PhantomData<(E, R)>,
+    phantom: PhantomData<(T, E, R)>,
 }
 
-impl<S, E, X> ApplicationBuilder<S, E, X> {
+impl<S, T, E, X> ApplicationBuilder<S, T, E, X>
+where
+    T: 'static,
+{
     /// Creates a new [ApplicationBuilder](struct.ApplicationBuilder.html) instance
     /// that wraps the initial_state. This is the more verbose way of initializing
     /// your application if you require specific configuration details to be changed
@@ -470,6 +497,7 @@ impl<S, E, X> ApplicationBuilder<S, E, X> {
         world.add_resource(pool);
         world.add_resource(EventChannel::<Event>::with_capacity(2000));
         world.add_resource(EventChannel::<UiEvent>::with_capacity(40));
+        world.add_resource(EventChannel::<TransEvent<T, StateEvent>>::with_capacity(2));
         world.add_resource(Errors::default());
         world.add_resource(FrameLimiter::default());
         world.add_resource(Stopwatch::default());
@@ -737,7 +765,7 @@ impl<S, E, X> ApplicationBuilder<S, E, X> {
     ///
     /// See the [example show for `ApplicationBuilder::new()`](struct.ApplicationBuilder.html#examples)
     /// for an example on how this method is used.
-    pub fn build<'a, T, I>(mut self, init: I) -> Result<CoreApplication<'a, T, E, X>>
+    pub fn build<'a, I>(mut self, init: I) -> Result<CoreApplication<'a, T, E, X>>
     where
         S: State<T, E> + 'a,
         I: DataInit<T>,
@@ -759,6 +787,10 @@ impl<S, E, X> ApplicationBuilder<S, E, X> {
             .world
             .exec(|mut ev: Write<EventChannel<Event>>| ev.register_reader());
 
+        let trans_reader_id = self
+            .world
+            .exec(|mut ev: Write<EventChannel<TransEvent<T, E>>>| ev.register_reader());
+
         Ok(CoreApplication {
             world: self.world,
             states: StateMachine::new(self.initial_state),
@@ -767,6 +799,7 @@ impl<S, E, X> ApplicationBuilder<S, E, X> {
             ignore_window_close: self.ignore_window_close,
             data,
             event_reader_id,
+            trans_reader_id,
         })
     }
 }