+++ src/prelude.rs
@@ -6,6 +6,8 @@ pub use {
     core::WithNamed,
     ecs::prelude::{Builder, World},
     game_data::{DataInit, GameData, GameDataBuilder},
-    state::{EmptyState, EmptyTrans, SimpleState, SimpleTrans, State, StateData, Trans},
+    state::{
+        EmptyState, EmptyTrans, SimpleState, SimpleTrans, State, StateData, Trans, TransEvent,
+    },
     state_event::StateEvent,
 };